backend/requirements.txt
Flask
flask-cors
openai
python-dotenv
PyPDF2
docx2txt
